test(observe::enrich): move test utilities to enrich_test_utils.rs
Extract common test utility functions and structs from enrich_tests.rs into a new file enrich_test_utils.rs and import them with a module declaration in enrich.rs. This improves code organization and reusability by separating test utilities from test cases.
